#38, #35, #5 Log out button with redirect
