+++ src/App.tsx
@@ -1,11 +1,13 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Route, Switch, BrowserRouter } from "react-router-dom";
 import Login from "./components/Login";
 import SignUp from "./components/SignUp";
-import { RecoilRoot } from "recoil";
+import { RecoilRoot, useRecoilValue, useSetRecoilState } from "recoil";
 import Dashboard from "./components/Dashboard";
 
 import { ThemeProvider, createMuiTheme } from "@material-ui/core/styles";
+import { auth } from "./firebase";
+import { currentUserState } from "./stateManagement/userAuth";
 
 const studentidrettTheme = createMuiTheme({
   palette: {
@@ -28,20 +30,30 @@ const studentidrettTheme = createMuiTheme({
 });
 
 function App() {
+  const setCurrentUser = useSetRecoilState(currentUserState);
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((user: any) => {
+      if (user != null) {
+        setCurrentUser(user.toJSON());
+        console.log("ikke null firebase", user);
+      } else {
+        setCurrentUser(null);
+      }
+    });
+    return unsubscribe;
+  }, []);
   return (
-    <RecoilRoot>
-      <ThemeProvider theme={studentidrettTheme}>
-        <div style={{ display: "flex", flexDirection: "row", padding: 20 }}>
-          <BrowserRouter>
-            <Switch>
-              <Route exact path="/login" component={Login} />
-              <Route exact path="/signup" component={SignUp} />
-              <Dashboard></Dashboard>
-            </Switch>
-          </BrowserRouter>
-        </div>
-      </ThemeProvider>
-    </RecoilRoot>
+    <ThemeProvider theme={studentidrettTheme}>
+      <div style={{ display: "flex", flexDirection: "row", padding: 20 }}>
+        <BrowserRouter>
+          <Switch>
+            <Route exact path="/login" component={Login} />
+            <Route exact path="/signup" component={SignUp} />
+            <Dashboard></Dashboard>
+          </Switch>
+        </BrowserRouter>
+      </div>
+    </ThemeProvider>
   );
 }
 export default App;
