+++ src/components/Dashboard.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Route } from "react-router";
 import { BrowserRouter, Switch } from "react-router-dom";
 import ChooseApplication from "./ChooseApplication";
@@ -8,21 +8,37 @@ import StudentCupForm from "./StudentCupForm";
 import StudentlekerForm from "./StudentlekerForm";
 import StudentNMForm from "./StudentNMForm";
 import UserProfile from "./UserProfile";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+
+import { currentUserState } from "../stateManagement/userAuth";
+import Login from "./Login";
+import SignUp from "./SignUp";
 
 export default function Dashboard() {
-  return (
-    <div style={{ display: "flex", flexDirection: "row", padding: 20 }}>
-      <BrowserRouter>
-        <DrawerBar />
-        <Switch>
+  const currentUser = useRecoilValue(currentUserState);
+
+  if (currentUser != null) {
+    return (
+      <div style={{ display: "flex", flexDirection: "row", padding: 20 }}>
+        <BrowserRouter>
+          <DrawerBar />
           <Route exact path="/" component={Home} />
           <Route exact path="/userprofile" component={UserProfile} />
           <Route exact path="/applications" component={ChooseApplication} />
           <Route path="/studentnm" component={StudentNMForm} />
           <Route path="/studentleker" component={StudentlekerForm} />
           <Route path="/studentcup" component={StudentCupForm} />
-        </Switch>
-      </BrowserRouter>
-    </div>
-  );
+        </BrowserRouter>
+      </div>
+    );
+  } else {
+    return (
+      <div style={{ display: "flex", flexDirection: "row", padding: 20 }}>
+        <BrowserRouter>
+          <Route exact path="/login" component={Login} />
+          <Route exact path="/signup" component={SignUp} />
+        </BrowserRouter>
+      </div>
+    );
+  }
 }
